Fix warnings

+++ Categories/NSData+Base64.m
@@ -72,7 +72,7 @@ void *NewBase64Decode(
 	size_t length,
 	size_t *outputLength)
 {
-	if (length == -1)
+	if ((NSInteger)length == -1)
 	{
 		length = strlen(inputBuffer);
 	}
